backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
langgraph>=0.2.0
langchain>=0.1.0
langchain-core>=0.1.0
sqlparse>=0.4.4
python-dotenv>=1.0.0
openai>=1.0.0
mistralai>=1.0.0
python-multipart>=0.0.6
